restructure: create "PostsList" in the file "feature" and create "NewsFeed" component to render it
