+++ src/featuers/posts/PostsList.js
@@ -1,21 +1,11 @@
 import React, { useState, useEffect } from "react";
-import { Box, Card, CircularProgress, styled } from "@mui/material";
+import { CircularProgress } from "@mui/material";
 import Post from "./Post";
-import Create from "../compose/Create";
+import { LoadingWrapper } from "./styles";
+
 import { useDispatch, useSelector } from "react-redux";
 import { fetchPosts } from "../../actions/postsAction";
 
-const LoadingWrapper = styled(Card)(({ theme }) => ({
-  elevation: 10,
-  display: "flex",
-  justifyContent: "center",
-  alignItems: "center",
-  marginLeft: theme.spacing(2),
-  marginRight: theme.spacing(2),
-  marginTop: theme.spacing(2),
-  maxWidth: 690,
-  height: "25vh",
-}));
 const NewsFeed = () => {
   const [currentId, setCurrentId] = useState(0);
   const dispatch = useDispatch();
@@ -28,19 +18,16 @@ const NewsFeed = () => {
   if (!posts) return null;
   if (isLoading) {
     return (
-      <Box flex={4} pt={1} sx={{ display: { sm: "block" } }}>
-        <Create currentId={currentId} setCurrentId={setCurrentId} />
+      <section>
         <LoadingWrapper>
           <CircularProgress />
         </LoadingWrapper>
-      </Box>
+      </section>
     );
   }
 
   return (
-    <Box flex={4} pt={1} sx={{ display: { sm: "block" } }}>
-      <Create currentId={currentId} setCurrentId={setCurrentId} />
-      {/* <UpdatePost currentId={currentId} setCurrentId={setCurrentId} /> */}
+    <section>
       {posts &&
         posts.length > 0 &&
         posts.map((post) => (
@@ -52,7 +39,7 @@ const NewsFeed = () => {
             />
           </div>
         ))}
-    </Box>
+    </section>
   );
 };
