+++ src/featuers/posts/styles.js
@@ -1,4 +1,4 @@
-import { styled, Button } from "@mui/material";
+import { styled, Button, Card } from "@mui/material";
 
 //AddForm
 // text input area
@@ -43,4 +43,17 @@ export const StyledButton = styled(Button)(({ theme }) => ({
   position: "absolute",
   bottom: "15px",
   right: "35px",
+}));
+
+//newsFeed
+const LoadingWrapper = styled(Card)(({ theme }) => ({
+  elevation: 10,
+  display: "flex",
+  justifyContent: "center",
+  alignItems: "center",
+  marginLeft: theme.spacing(2),
+  marginRight: theme.spacing(2),
+  marginTop: theme.spacing(2),
+  maxWidth: 690,
+  height: "25vh",
 }));
